requirements.txt
requests==2.27.1
python-dateutil==2.8.2
chardet==4.0.0